pyproject.toml
[project]
name = "spotify-statistics"
version = "0.1.0"
description = "Create statistics from your spotify listening history"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "ruff",
    "requests",
    "python-dotenv",
    "tqdm",
]

[tool.ruff]
lint.select = [
    "F",        # Pyflakes
    "UP",       # pyupgrade
    "B018",     # Found useless expression.
    "I",        # isort
    "RUF",      # ruff
]
line-length = 110

[tool.ruff.format]
quote-style = "double"
indent-style = "space"